fix(suite-native): staking navigation fix
