+++ suite-native/module-earn/src/hooks/useEarnReviewBackNavigation.ts
@@ -1,4 +1,4 @@
-import { useCallback } from 'react';
+import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { useNavigation } from '@react-navigation/native';
@@ -6,14 +6,13 @@ import { useNavigation } from '@react-navigation/native';
 import { formDraftActions, sendFormActions } from '@suite-common/wallet-core';
 import { type AccountKey } from '@suite-common/wallet-types';
 import { getFormDraftKey } from '@suite-common/wallet-utils';
-import { useOverrideBackNavigation } from '@suite-native/navigation';
 import {
     type TransactionReviewOutputsState,
     selectIsTransactionReviewInProgress,
     useShowReviewCancellationAlert,
 } from '@suite-native/transaction-management';
 
-type EarnReviewFormType = 'stake' | 'claim';
+type EarnReviewFormType = 'stake' | 'unstake' | 'claim';
 
 export const useEarnReviewBackNavigation = (
     formType: EarnReviewFormType,
@@ -27,21 +26,27 @@ export const useEarnReviewBackNavigation = (
     const navigation = useNavigation();
     const showReviewCancellationAlert = useShowReviewCancellationAlert();
 
-    const onNavigateBack = useCallback(async () => {
-        if (isTransactionReviewInProgress) {
-            const { wasReviewCanceled } = await showReviewCancellationAlert();
-            if (!wasReviewCanceled) return;
-        }
-        dispatch(sendFormActions.discardTransaction());
-        dispatch(formDraftActions.removeDraft({ key: getFormDraftKey(formType, '') }));
-        navigation.goBack();
+    useEffect(() => {
+        const unsubscribe = navigation.addListener('beforeRemove', e => {
+            if (e.data.action.type === 'GO_BACK' && isTransactionReviewInProgress) {
+                e.preventDefault();
+                showReviewCancellationAlert().then(({ wasReviewCanceled }) => {
+                    if (wasReviewCanceled) navigation.dispatch(e.data.action);
+                });
+
+                return;
+            }
+
+            dispatch(sendFormActions.discardTransaction());
+            dispatch(formDraftActions.removeDraft({ key: getFormDraftKey(formType, '') }));
+        });
+
+        return unsubscribe;
     }, [
+        navigation,
         isTransactionReviewInProgress,
         showReviewCancellationAlert,
         dispatch,
-        navigation,
         formType,
     ]);
-
-    useOverrideBackNavigation({ onNavigateBack });
 };
