+++ suite-native/module-earn/src/hooks/useHandleOnClaimTransactionReview.ts
@@ -1,25 +1,19 @@
-import { useCallback, useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useCallback } from 'react';
+import { useDispatch } from 'react-redux';
 
 import { useNavigation } from '@react-navigation/native';
 import { isFulfilled, isRejected } from '@reduxjs/toolkit';
 
-import { formDraftActions, sendFormActions } from '@suite-common/wallet-core';
 import { type AccountKey } from '@suite-common/wallet-types';
-import { getFormDraftKey } from '@suite-common/wallet-utils';
 import {
     type RootStackParamList,
     type RootStackRoutes,
     type StackNavigationProps,
 } from '@suite-native/navigation';
 import { signEthClaimTransactionNativeThunk } from '@suite-native/staking';
-import {
-    type TransactionReviewOutputsState,
-    selectIsTransactionReviewInProgress,
-    useShowReviewCancellationAlert,
-} from '@suite-native/transaction-management';
 
 import { handleEarnReviewError } from '../utils';
+import { useEarnReviewBackNavigation } from './useEarnReviewBackNavigation';
 import { useShowDeviceDisconnectedDuringEarnReviewAlert } from './useShowDeviceDisconnectedDuringEarnReviewAlert';
 import { useShowPushTransactionFailedDuringReviewAlert } from './useShowPushTransactionFailedDuringReviewAlert';
 
@@ -37,32 +31,13 @@ export const useHandleOnClaimTransactionReview = ({
     accountKey,
     onTransactionSubmitted,
 }: HandleOnClaimTransactionReviewProps) => {
-    const isTransactionReviewInProgress = useSelector((state: TransactionReviewOutputsState) =>
-        selectIsTransactionReviewInProgress(state, 'claim', accountKey),
-    );
+    useEarnReviewBackNavigation('claim', accountKey);
 
     const dispatch = useDispatch();
     const navigation = useNavigation<NavigationProps>();
-    const showReviewCancellationAlert = useShowReviewCancellationAlert();
     const showDeviceDisconnectedAlert = useShowDeviceDisconnectedDuringEarnReviewAlert();
     const { showPushTransactionFailedAlert, showPendingTransactionConflictAlert } =
         useShowPushTransactionFailedDuringReviewAlert('claim');
-
-    useEffect(() => {
-        const unsubscribe = navigation.addListener('beforeRemove', e => {
-            if (e.data.action.type === 'GO_BACK' && isTransactionReviewInProgress) {
-                e.preventDefault();
-                showReviewCancellationAlert();
-
-                return;
-            }
-
-            dispatch(sendFormActions.discardTransaction());
-            dispatch(formDraftActions.removeDraft({ key: getFormDraftKey('claim', '') }));
-        });
-
-        return unsubscribe;
-    }, [navigation, isTransactionReviewInProgress, showReviewCancellationAlert, dispatch]);
 
     const handleOnClaimTransactionReview = useCallback(async () => {
         const response = await dispatch(signEthClaimTransactionNativeThunk({ accountKey }));
