+++ suite-native/module-earn/src/hooks/useHandleOnUnstakeTransactionReview.ts
@@ -1,25 +1,18 @@
 import { useCallback } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 
 import { useNavigation } from '@react-navigation/native';
 import { isFulfilled, isRejected } from '@reduxjs/toolkit';
 
-import { formDraftActions, sendFormActions } from '@suite-common/wallet-core';
 import { type AccountKey } from '@suite-common/wallet-types';
-import { getFormDraftKey } from '@suite-common/wallet-utils';
 import {
     type RootStackParamList,
     type RootStackRoutes,
     type StackNavigationProps,
-    useOverrideBackNavigation,
 } from '@suite-native/navigation';
 import { signEthUnstakeTransactionNativeThunk } from '@suite-native/staking';
-import {
-    type TransactionReviewOutputsState,
-    selectIsTransactionReviewInProgress,
-    useShowReviewCancellationAlert,
-} from '@suite-native/transaction-management';
 
+import { useEarnReviewBackNavigation } from './useEarnReviewBackNavigation';
 import { useShowDeviceDisconnectedDuringEarnReviewAlert } from './useShowDeviceDisconnectedDuringEarnReviewAlert';
 import { useShowPushTransactionFailedDuringUnstakeReviewAlert } from './useShowPushTransactionFailedDuringUnstakeReviewAlert';
 
@@ -45,28 +38,13 @@ export const useHandleOnUnstakeTransactionReview = ({
     amount,
     onTransactionSubmitted,
 }: HandleOnUnstakeTransactionReviewProps) => {
-    const isTransactionReviewInProgress = useSelector((state: TransactionReviewOutputsState) =>
-        selectIsTransactionReviewInProgress(state, 'unstake', accountKey),
-    );
+    useEarnReviewBackNavigation('unstake', accountKey);
 
     const dispatch = useDispatch();
     const navigation = useNavigation<NavigationProps>();
-    const showReviewCancellationAlert = useShowReviewCancellationAlert();
     const showDeviceDisconnectedAlert = useShowDeviceDisconnectedDuringEarnReviewAlert();
     const { showPushTransactionFailedAlert, showPendingTransactionConflictAlert } =
         useShowPushTransactionFailedDuringUnstakeReviewAlert();
-
-    const onNavigateBack = useCallback(async () => {
-        if (isTransactionReviewInProgress) {
-            const { wasReviewCanceled } = await showReviewCancellationAlert();
-            if (!wasReviewCanceled) return;
-        }
-        dispatch(sendFormActions.discardTransaction());
-        dispatch(formDraftActions.removeDraft({ key: getFormDraftKey('unstake', '') }));
-        navigation.goBack();
-    }, [isTransactionReviewInProgress, showReviewCancellationAlert, dispatch, navigation]);
-
-    useOverrideBackNavigation({ onNavigateBack });
 
     const handleOnUnstakeTransactionReview = useCallback(async () => {
         const response = await dispatch(
